ref is uint32

+++ include/ttypt/rec.h
@@ -3,11 +3,15 @@
  * @brief Recall kernel: uniform candidate sets + a generic ranking loop.
  *
  * Pure C, no domain math. Candidate sets are sorted, deduplicated lists of
- * 64-bit refs (merge-join, no hashing). The ranking loop is a bounded,
- * streaming min-heap: an axis streams refs + scores straight into
- * rec_rank_push; a set is materialized only when a second axis needs to
- * join. Weighted composition across axes lives in the consumer's score
- * function.
+ * 32-bit refs (merge-join, no hashing). A ref IS a qmap record reference
+ * (e.g. the id returned by qmap_put() under QM_AINDEX) — the same
+ * primary-key space every qmap-backed store shares, never an axis's own
+ * internal key. An axis may index by whatever it needs internally (e.g.
+ * libislet's 64-bit morton keys); those internal keys are not refs and
+ * never leave the axis. The ranking loop is a bounded, streaming min-heap:
+ * an axis streams refs + scores straight into rec_rank_push; a set is
+ * materialized only when a second axis needs to join. Weighted composition
+ * across axes lives in the consumer's score function.
  *
  * Kernel is optional and additive: raw qmap / domain entry points in the
  * axis libraries are untouched. Ref mapping is the consumer's job — the
@@ -34,7 +38,7 @@
 extern "C" {
 #endif
 
-typedef uint64_t rec_ref_t;
+typedef uint32_t rec_ref_t;
 
 /** @defgroup rec_set Candidate sets
  *  @{
@@ -52,7 +56,7 @@ void rec_set_push(rec_set_t *s, rec_ref_t r);
 void rec_set_seal(rec_set_t *s);
 
 /** Drain a qmap handle's iteration into the set (keys must be fixed-length,
- *  <= 8 bytes; refs are read from the key bytes). Excludes nothing; call
+ *  <= 4 bytes; refs are read from the key bytes). Excludes nothing; call
  *  rec_set_seal afterwards. Returns 0 on success, -1 if keys are variable
  *  length or wider than a ref. */
 int rec_set_fill_qmap_iter(rec_set_t *s, uint32_t hd);
